window frozen check modify. Co-Authored-By: Agent

+++ wm/src/window_session_impl.cpp
@@ -8565,7 +8565,7 @@ void WindowSessionImpl::AddSetUIContentTimeoutCheck()
             }
         }
         SingletonContainer::Get<WindowInfoReporter>().ReportWindowFrozen(
-            static_cast<int32_t>(WindowDFXHelperType::WINDOW_UICONTENT_TIMEOUT_CHECK), oss.str());
+            WindowDFXHelperType::WINDOW_UICONTENT_TIMEOUT_CHECK, oss.str());
 
         if (WindowHelper::IsUIExtensionWindow(window->GetType())) {
             window->NotifyExtensionTimeout(TimeoutErrorCode::SET_UICONTENT_TIMEOUT);
@@ -8640,7 +8640,7 @@ void WindowSessionImpl::AddSetUIExtensionDestroyTimeoutCheck()
             }
         }
         SingletonContainer::Get<WindowInfoReporter>().ReportWindowFrozen(
-            static_cast<int32_t>(WindowDFXHelperType::WINDOW_UIEXT_DESTROY_TIMEOUT_CHECK), oss.str());
+            WindowDFXHelperType::WINDOW_UIEXT_DESTROY_TIMEOUT_CHECK, oss.str());
 
         if (WindowHelper::IsUIExtensionWindow(window->GetType())) {
             window->NotifyExtensionTimeout(TimeoutErrorCode::SET_UIEXTENSION_DESTROY_TIMEOUT);
